refactor(renderer-build): split runtime-source.ts into plugins/
runtime-source.ts(217줄)는 한 파일 안에 (1) SDK alias plugin과 SDK
소스 미러, (2) react/scheduler 사이드카 resolve, (3) JSX pragma 주입
plugin 세 책임이 섞여 있었다. 책임 단위로 세 파일로 분리한다.

- plugins/sdk-shim.ts: SDK_NAMESPACE, RENDERER_CORE_SOURCE,
  RENDERER_REACT_SOURCE, SDK_SOURCES, createRendererRuntimePlugin.
- plugins/host-runtime.ts: HOST_RUNTIME_SIDE_CARS, resolveHostRuntimePath,
  resolveRendererRuntimeDependency, loadTransformedRuntimeSource.
- plugins/source-loader.ts: createRendererSourcePlugin (JSX pragma 주입).
- plugins/index.ts: 두 plugin factory 재export.

builder.test.ts의 SDK ↔ 원본 equivalence test 16/16 pass로 분리 후
동작 동일 검증.

issue #144 Phase B3.
