+++ packages/renderer-build/src/plugins/index.ts
@@ -1,0 +1,2 @@
+export { createRendererRuntimePlugin } from "./sdk-shim.js";
+export { createRendererSourcePlugin } from "./source-loader.js";
